infra: cleanups
