+++ packages/legend-application-data-cube-bootstrap/package.json
@@ -48,17 +48,17 @@
     "@finos/legend-application-data-cube": "workspace:*",
     "@finos/legend-extension-dsl-data-space": "workspace:*",
     "@finos/legend-shared": "workspace:*",
-    "@types/react": "18.3.12",
-    "react": "18.3.1"
+    "@types/react": "19.0.1",
+    "react": "19.0.0"
   },
   "devDependencies": {
     "@finos/legend-dev-utils": "workspace:*",
     "cross-env": "7.0.3",
-    "eslint": "9.15.0",
+    "eslint": "9.17.0",
     "npm-run-all": "4.1.5",
     "rimraf": "6.0.1",
-    "sass": "1.81.0",
-    "typescript": "5.6.3"
+    "sass": "1.83.0",
+    "typescript": "5.7.2"
   },
   "publishConfig": {
     "directory": "build/publishContent"
