+++ packages/legend-application-data-cube-deployment/docker/entrypoint.sh
@@ -4,7 +4,7 @@ PROTOCOL=${PROTOCOL:-http}
 ENGINE_HOST=${ENGINE_HOST:-localhost}
 ENGINE_PORT=${ENGINE_PORT:-6300}
 DEPOT_HOST=${DEPOT_HOST:-localhost}
-DEPOT_PORT=${DEPOT_PORT:-7000}
+DEPOT_PORT=${DEPOT_PORT:-6200}
 
 sed -i 's~${PROTOCOL}~'$PROTOCOL'~g' /config/config.json
 sed -i 's~${ENGINE_HOST}~'$ENGINE_HOST'~g' /config/config.json
