+++ packages/legend-application-data-cube-deployment/package.json
@@ -20,7 +20,7 @@
     "build:tailwindcss": "tailwindcss -i ./style/tailwind.css -o ./lib/tailwind.css --minify",
     "build:webpack": "cross-env NODE_ENV=production webpack --mode production",
     "build-dry:docker": "../../scripts/docker/dry-build-docker-image.sh legend-data-cube ./dist/data-cube",
-    "bundle": "yarn clean && yarn build:webpack && node ./scripts/setup.js ./dist/data-cube ",
+    "bundle": "yarn clean && yarn build:webpack && node ./scripts/setup.js ./dist/data-cube",
     "bundle:analyze": "cross-env NODE_ENV=production webpack --mode production --analyze",
     "bundle:fast": "cross-env NODE_ENV=production PRODUCTION_MODE=fast webpack --mode production && node ./scripts/setup.js ./dist/data-cube",
     "clean": "yarn clean:cache && rimraf \"dist\"",
@@ -42,14 +42,14 @@
     "@finos/legend-dev-utils": "workspace:*",
     "copy-webpack-plugin": "12.0.2",
     "cross-env": "7.0.3",
-    "eslint": "9.15.0",
+    "eslint": "9.17.0",
     "npm-run-all": "4.1.5",
     "rimraf": "6.0.1",
-    "tailwindcss": "3.4.15",
-    "typescript": "5.6.3",
-    "webpack": "5.95.0",
+    "tailwindcss": "3.4.16",
+    "typescript": "5.7.2",
+    "webpack": "5.97.1",
     "webpack-bundle-analyzer": "4.10.2",
     "webpack-cli": "5.1.4",
-    "webpack-dev-server": "5.1.0"
+    "webpack-dev-server": "5.2.0"
   }
 }
