+++ packages/legend-application-data-cube/package.json
@@ -51,23 +51,23 @@
     "@finos/legend-server-depot": "workspace:*",
     "@finos/legend-shared": "workspace:*",
     "@finos/legend-storage": "workspace:*",
-    "@types/react": "18.3.12",
-    "@types/react-dom": "18.3.1",
+    "@types/react": "19.0.1",
+    "@types/react-dom": "19.0.2",
     "mobx": "6.13.5",
-    "mobx-react-lite": "4.0.7",
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "mobx-react-lite": "4.1.0",
+    "react": "19.0.0",
+    "react-dom": "19.0.0"
   },
   "devDependencies": {
     "@finos/legend-dev-utils": "workspace:*",
     "@jest/globals": "29.7.0",
     "cross-env": "7.0.3",
-    "eslint": "9.15.0",
+    "eslint": "9.17.0",
     "jest": "29.7.0",
     "npm-run-all": "4.1.5",
     "rimraf": "6.0.1",
-    "sass": "1.81.0",
-    "typescript": "5.6.3"
+    "sass": "1.83.0",
+    "typescript": "5.7.2"
   },
   "peerDependencies": {
     "react": "^18.0.0"
