+++ packages/legend-application-studio-deployment/docker/entrypoint.sh
@@ -6,7 +6,7 @@ ENGINE_PORT=${ENGINE_PORT:-6300}
 SDLC_HOST=${SDLC_HOST:-localhost}
 SDLC_PORT=${SDLC_PORT:-6100}
 DEPOT_HOST=${DEPOT_HOST:-localhost}
-DEPOT_PORT=${DEPOT_PORT:-7000}
+DEPOT_PORT=${DEPOT_PORT:-6200}
 
 sed -i 's~${PROTOCOL}~'$PROTOCOL'~g' /config/config.json
 sed -i 's~${ENGINE_HOST}~'$ENGINE_HOST'~g' /config/config.json
