+++ packages/legend-art/package.json
@@ -58,8 +58,8 @@
     "@fontsource/roboto-mono": "5.1.0",
     "@fontsource/roboto-serif": "5.1.1",
     "@fontsource/ubuntu-mono": "5.1.0",
-    "@mui/material": "6.2.0",
-    "@mui/system": "6.2.0",
+    "@mui/material": "6.2.1",
+    "@mui/system": "6.2.1",
     "@mui/x-date-pickers": "7.23.2",
     "@types/react": "19.0.1",
     "@types/react-window": "1.8.8",
@@ -81,7 +81,7 @@
     "react-resize-detector": "12.0.1",
     "react-rnd": "10.4.13",
     "react-select": "5.9.0",
-    "react-window": "1.8.10",
+    "react-window": "1.8.11",
     "remark-gfm": "4.0.0",
     "tailwind-merge": "2.5.5"
   },
