+++ packages/legend-art/src/index.ts
@@ -74,7 +74,6 @@ export * from './notification/Notification.js';
 export * from './utils/Backdrop.js';
 export * from './utils/ClickAwayListener.js';
 export * from './utils/LegendStyleProvider.js';
-export * from './utils/ComponentsConfiguration.js';
 export * from './utils/ComponentUtils.js';
 
 export * from './markdown/MarkdownTextViewer.js';
