+++ packages/legend-server-showcase-deployment/Dockerfile
@@ -1,4 +1,4 @@
-FROM node:16-bullseye-slim
+FROM node:bullseye-slim
 COPY dist /app/showcase-server/
 
 CMD node /app/showcase-server/index.mjs /app/showcase-server/config.json
