+++ packages/legend-vscode-extension-dependencies/package.json
@@ -70,7 +70,7 @@
     "rollup-plugin-flow": "1.1.1",
     "rollup-plugin-import-css": "3.5.8",
     "rollup-plugin-postcss": "4.0.2",
-    "tailwindcss": "3.4.15",
+    "tailwindcss": "3.4.16",
     "typescript": "5.7.2"
   },
   "publishConfig": {
